+++ lizard_krw/views.py
@@ -374,6 +374,7 @@ def select_area(request, template='lizard_krw/select_area.html',
          'workspaces': {'user': [special_homepage_workspace]},
          'javascript_hover_handler': 'popup_hover_handler',
          'javascript_click_handler': 'homepage_area_click_handler',
+         'use_workspaces': False,
          'crumbs': crumbs},
         context_instance=RequestContext(request))
 
@@ -385,11 +386,6 @@ def krw_browser(request, template='lizard_krw/krw-browser.html',
     Automatically makes new workspace if not yet available
 
     """
-    workspace_manager = WorkspaceManager(request)
-    workspaces = workspace_manager.load_or_create()
-    date_range_form = DateRangeForm(
-        current_start_end_dates(request, for_form=True))
-
     if crumbs_prepend is not None:
         crumbs = crumbs_prepend
     else:
@@ -400,9 +396,7 @@ def krw_browser(request, template='lizard_krw/krw-browser.html',
 
     return render_to_response(
         template,
-        {'date_range_form': date_range_form,
-         'workspaces': workspaces,
-         'javascript_hover_handler': 'popup_hover_handler',
+        {'javascript_hover_handler': 'popup_hover_handler',
          'javascript_click_handler': 'popup_click_handler',
          'crumbs': crumbs},
         context_instance=RequestContext(request))